Data preperation
I preprocess an audio dataset and get it ready for music genre classification. Specifically, I implement code to batch process the Kaggle music dataset, in order to extract MFCCs and genre labels. I save the data in a json file, in a format that's convenient for retrieval when training the classifier.